Added Firefox
This patch adds firefox to our testing grid. It requires adding
the karma-firefox-launcher to our karma configuration, as well as
switching protractor's 'directConnect' feature from true to false.
This will trigger protractor to start up its own selenium standalone
server, from which the browsers are controlled.

To speed things up, this patch also ups the concurrency of both
harnesses to 2, so tests are run in parallel.