Switch to using selenium server started by webdriver-manager
This patch updates the protractor and karma configuration to talk
to the selenium server directly. In Karma's case, we switch to using
the webdriver-launcher, and provide it with the address of our selenium
grid. In Protractor's case, we only have to do the latter. Notice that
when running protractor, it no longer attempts to start selenium itself.