Bind all servers to external IP's.
This is actually a security risk, so proceed with caution.

In order for a 'remote' selenium grid - such as one hosted in a docker
VM or on a third party service - to be able to run tests on your
system, it will need to be able to reach your system over its network.

In the simplest case, this means that our application server will need
to bind to the public interface (0.0.0.0), or a detected local IP Address.
Only the latter of these two will work for Karma, as it will pass it's
bound hostname to selenium; only a public IP will do.