Switched from using webdriver-manager to selenium-cli
With this change, the 'selenium' target now launches a docker-hosted
selenium grid with two nodes: firefox and chrome. Since the docker
environment is effectively network-isolated from the host, it accurately
simulates a remote selenium grid, as the only way for its browser to
access our testing host is via a public IP address.